Update ipahealthcheck_exporter.go
Fixed issues with file permissions when using sudo

+++ ipahealthcheck_exporter.go
@@ -116,6 +116,9 @@ func (ic ipahealthcheckCollector) Collect(ch chan<- prometheus.Metric) {
 
 	var checks []ipaCheck
 	tmpFile, err := os.CreateTemp("/dev/shm", "ipa-healthcheck.out")
+	if sudo {
+		cmd := exec.Command("sudo chown root", tmpFile.Name())
+	}
 	if err != nil {
 		log.Fatal("Cannot write ipa-healthcheck output for parsing: ", err)
 	}
@@ -224,7 +227,11 @@ func (ic ipahealthcheckCollector) Collect(ch chan<- prometheus.Metric) {
 		}
 	}
 
-	defer os.Remove(tmpFile.Name())
+	if sudo {
+		cmd := exec.Command("sudo rm", tmpFile.Name())
+	} else {
+		defer os.Remove(tmpFile.Name())
+	}
 }
 
 func main() {
